+++ crates/ra_mbe/Cargo.toml
@@ -5,6 +5,7 @@ version = "0.1.0"
 authors = ["Aleksey Kladov <aleksey.kladov@gmail.com>"]
 
 [dependencies]
+ra_syntax = { path = "../ra_syntax" }
 tt = { path = "../ra_tt", package = "ra_tt" }
+
 rustc-hash = "1.0.0"
-smol_str = "0.1.9"
