+++ crates/ra_mbe/src/lib.rs
@@ -1,3 +1,8 @@
+/// `mbe` (short for Macro By Example) crate contains code for handling
+/// `macro_rules` macros. It uses `TokenTree` (from `ra_tt` package) as the
+/// interface, although it contains some code to bridge `SyntaxNode`s and
+/// `TokenTree`s as well!
+
 macro_rules! impl_froms {
     ($e:ident: $($v:ident), *) => {
         $(
@@ -13,14 +18,16 @@ macro_rules! impl_froms {
 mod tt_cursor;
 mod mbe_parser;
 mod mbe_expander;
+mod syntax_bridge;
 
-use smol_str::SmolStr;
+use ra_syntax::SmolStr;
 
 pub use tt::{Delimiter, Punct};
 
 pub use crate::{
     mbe_parser::parse,
     mbe_expander::exapnd,
+    syntax_bridge::macro_call_to_tt,
 };
 
 #[derive(Debug)]
