+++ crates/ra_tt/Cargo.toml
@@ -5,5 +5,4 @@ version = "0.1.0"
 authors = ["Aleksey Kladov <aleksey.kladov@gmail.com>"]
 
 [dependencies]
-rustc-hash = "1.0.0"
 smol_str = "0.1.9"
